opetmize path map insertion
